app/config.py
# app/config.py
from pydantic_settings import BaseSettings
import os
from dotenv import load_dotenv

load_dotenv()  # load environment variables from .env file

class Settings(BaseSettings):
    SECRET_KEY: str = os.getenv("SECRET_KEY", "your-super-secret-key-for-development")
    ALGORITHM: str = "HS256"
    ACCESS_TOKEN_EXPIRE_MINUTES: int = 30
    
    # Database
    DATABASE_URL: str = os.getenv("DATABASE_URL", "sqlite:///./sql_app.db")
    
    # Email
    SMTP_HOST: str = os.getenv("SMTP_HOST", "smtp.gmail.com")
    SMTP_PORT: int = int(os.getenv("SMTP_PORT", "587"))
    SMTP_USER: str = os.getenv("SMTP_USER", "your-email@gmail.com")
    SMTP_PASSWORD: str = os.getenv("SMTP_PASSWORD", "your-password")
    EMAIL_FROM: str = os.getenv("EMAIL_FROM", "your-email@gmail.com")
    
    # Application
    BASE_URL: str = os.getenv("BASE_URL", "http://localhost:8000")
    
    # File uploads
    MAX_UPLOAD_SIZE: int = 5 * 1024 * 1024  # 5MB

settings = Settings()