requirements.txt
# requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
itsdangerous==2.1.2
cryptography==41.0.4
python-dotenv==1.0.0
email-validator==2.0.0
pytest==7.4.3
httpx==0.25.0
psycopg2-binary